requirements.txt
Flask==3.0.3
requests==2.31.0
python-dotenv==1.0.1
twilio==9.0.0
